+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Fixed log for generating key size
+- Fixed issues with getting correct filename, line number and function name
 
 ## [1.0.0-beta.6]
 ### Fixed